// src/types/r3eTypes.ts
export interface ITireData<T> {
	FrontLeft: T;
	FrontRight: T;
	RearLeft: T;
	RearRight: T;
}